Fix for NR gamma, NR Ldetail history message and History NR decimal values

+++ rtengine/procparams.h
@@ -244,7 +244,7 @@ class WBParams {
         double	luma;
         double	Ldetail;
         double  chroma;
-		float	gamma;
+		double	gamma;
 		double	expcomp;
 	};
